backend/requirements.txt
googletrans
googlemaps
twilio
flask
google
sports.py
gunicorn==19.9.0